+++ src/main/java/org/micromanager/lightsheetmanager/model/devices/cameras/PVCamera.java
@@ -91,7 +91,7 @@ public class PVCamera extends CameraBase implements LightSheetCamera {
     public double getRowReadoutTime() {
         Rectangle roi = getROI();
         if (hasProperty(Properties.READOUT_TIME)) {
-            final float readoutTimeMs = getPropertyFloat(Properties.READOUT_TIME) / 1e6f;
+            final double readoutTimeMs = getPropertyFloat(Properties.READOUT_TIME) / 1e6;
             return (readoutTimeMs / roi.height);
         } else {
             return 0.01;  // TODO get more accurate value
@@ -99,32 +99,32 @@ public class PVCamera extends CameraBase implements LightSheetCamera {
     }
 
     @Override
-    public float getReadoutTime(CameraMode cameraMode) {
-        float readoutTimeMs = 10.0f;
+    public double getReadoutTime(CameraMode cameraMode) {
+        double readoutTimeMs = 10.0;
         switch (cameraMode) {
             case OVERLAP:
-                readoutTimeMs = 0.0f;
+                readoutTimeMs = 0.0;
                 break;
             case PSEUDO_OVERLAP:
                 if (isKinetix() || isPrime95B()) {
-                    final float preTime = getPropertyFloat(Properties.PRE_TRIGGER_TIME);
-                    readoutTimeMs = preTime / 1e6f;
+                    final double preTime = getPropertyFloat(Properties.PRE_TRIGGER_TIME);
+                    readoutTimeMs = preTime / 1e6;
                     // for safety we make sure to wait at least a quarter millisecond to trigger
                     //   (may have hidden assumptions in other code about at least one tic wait)
-                    if (readoutTimeMs < 0.249f) {
-                        readoutTimeMs = 0.25f;
+                    if (readoutTimeMs < 0.249) {
+                        readoutTimeMs = 0.25;
                     }
                 } else { // original Prime
-                    readoutTimeMs = 0.25f;
+                    readoutTimeMs = 0.25;
                 }
                 break;
             case VIRTUAL_SLIT:
-                readoutTimeMs = getPropertyFloat(Properties.READOUT_TIME) / 1e6f;
+                readoutTimeMs = getPropertyFloat(Properties.READOUT_TIME) / 1e6;
                 break;
             case EDGE: // fall through to next case
             case LEVEL:
-                final float readoutTime = getPropertyFloat(Properties.READOUT_TIME);
-                final float endGlobalToTrig = getPropertyFloat(Properties.PRE_TRIGGER_TIME) + 2 * readoutTime;
+                final double readoutTime = getPropertyFloat(Properties.READOUT_TIME);
+                final double endGlobalToTrig = getPropertyFloat(Properties.PRE_TRIGGER_TIME) + 2 * readoutTime;
                 // this factor of 2 is empirical 08-Jan-2021; I'm not sure why it's needed but that is the missing piece it seems
                 readoutTimeMs = endGlobalToTrig / 1e6f;
                 break;
@@ -135,22 +135,22 @@ public class PVCamera extends CameraBase implements LightSheetCamera {
     }
 
     @Override
-    public float getResetTime(CameraMode cameraMode) {
-        float resetTimeMs;
+    public double getResetTime(CameraMode cameraMode) {
+        double resetTimeMs;
         if (cameraMode == CameraMode.VIRTUAL_SLIT) {
-            resetTimeMs = 0.0f;
+            resetTimeMs = 0.0;
         } else {
             // TODO(Jon): Confirm that the Kinetix camera is like the Prime 95B
 
             // Photometrics Prime 95B is very different from other cameras so handle it as special case
             if (isKinetix() || isPrime95B()) {
-                final float trigToGlobal = getPropertyFloat(Properties.POST_TRIGGER_TIME)
+                final double trigToGlobal = getPropertyFloat(Properties.POST_TRIGGER_TIME)
                         + getPropertyFloat(Properties.READOUT_TIME);
                 // it appears as of end-May 2017 that the clearing time is actually rolled into the post-trigger
                 //    time despite Photometrics documentation to the contrary
-                resetTimeMs = trigToGlobal / 1e6f;
+                resetTimeMs = trigToGlobal / 1e6;
             } else {
-                resetTimeMs = 14.25f;  // strange number just to make it easy to find later; I think the original Prime needs to be added
+                resetTimeMs = 14.25;  // strange number just to make it easy to find later; I think the original Prime needs to be added
             }
         }
         return resetTimeMs;
